Move WaitFor functions to own file
